requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-cors==3.0.10
pywebview==3.7.2
pyinstaller==6.13.0 